delete unused PWM IOCTLs
