+++ ROMFS/px4fmu_common/init.d-posix/airframes/1040_standard_vtol
@@ -76,8 +76,6 @@ param set-default NAV_ACC_RAD 5
 
 param set-default VT_FWD_THRUST_EN 4
 param set-default VT_F_TRANS_THR 0.75
-param set-default VT_MOT_ID 1234
-param set-default VT_FW_MOT_OFFID 1234
 param set-default VT_B_TRANS_DUR 8
 param set-default VT_TYPE 2
