+++ ROMFS/px4fmu_common/init.d/airframes/1002_standard_vtol.hil
@@ -55,8 +55,6 @@ param set-default RTL_RETURN_ALT 30
 param set-default SDLOG_DIRS_MAX 7
 
 param set-default VT_F_TRANS_THR 0.75
-param set-default VT_MOT_ID 1234
-param set-default VT_FW_MOT_OFFID 1234
 param set-default VT_TYPE 2
 
 param set-default CA_AIRFRAME 2
