+++ ROMFS/px4fmu_common/init.d/airframes/13013_deltaquad
@@ -116,8 +116,6 @@ param set-default MAV_1_FORWARD 1
 param set-default SER_TEL2_BAUD 57600
 
 param set-default VT_TYPE 2
-param set-default VT_MOT_ID 1234
-param set-default VT_FW_MOT_OFFID 1234
 param set-default VT_F_TRANS_THR 1
 param set-default VT_PITCH_MIN 8
 param set-default VT_FW_QC_P 55
@@ -126,7 +124,6 @@ param set-default VT_TRANS_MIN_TM 15
 param set-default VT_B_TRANS_DUR 8
 param set-default VT_FWD_THRUST_SC 4
 param set-default VT_F_TRANS_DUR 1
-param set-default VT_IDLE_PWM_MC 1025
 param set-default VT_B_REV_OUT 0.5
 param set-default VT_B_TRANS_THR 0.7
 param set-default VT_TRANS_TIMEOUT 22
