+++ ROMFS/px4fmu_common/init.d/airframes/13014_vtol_babyshark
@@ -86,10 +86,7 @@ param set-default VT_B_DEC_MSS 1.5
 param set-default VT_B_TRANS_DUR 12
 param set-default VT_ELEV_MC_LOCK 0
 param set-default VT_FWD_THRUST_SC 1.2
-param set-default VT_FW_MOT_OFFID 5678
 param set-default VT_F_TR_OL_TM 8
-param set-default VT_IDLE_PWM_MC 1000
-param set-default VT_MOT_ID 5678
 param set-default VT_PSHER_RMP_DT 2
 param set-default VT_TRANS_MIN_TM 4
 param set-default VT_TYPE 2
