+++ platforms/nuttx/src/px4/nxp/kinetis/led_pwm/led_pwm.cpp
@@ -278,15 +278,15 @@ led_pwm_servo_set(unsigned channel, uint8_t  cvalue)
 
 	return 0;
 }
-unsigned
-led_pwm_servo_get(unsigned channel)
+
+unsigned led_pwm_servo_get(unsigned channel)
 {
 	if (channel >= 3) {
 		return 0;
 	}
 
 	unsigned timer = led_pwm_channels[channel].timer_index;
-	servo_position_t value = 0;
+	uint16_t value = 0;
 
 	/* test timer for validity */
 	if ((led_pwm_timers[timer].base == 0) ||
@@ -298,8 +298,8 @@ led_pwm_servo_get(unsigned channel)
 	unsigned period = led_pwm_timer_get_period(timer);
 	return ((value + 1) * 255 / period);
 }
-int
-led_pwm_servo_init(void)
+
+int led_pwm_servo_init()
 {
 	/* do basic timer initialisation first */
 	for (unsigned i = 0; i < arraySize(led_pwm_timers); i++) {
