+++ platforms/nuttx/src/px4/stm/stm32_common/dshot/dshot.c
@@ -43,7 +43,7 @@
 #include <stm32_tim.h>
 #include <px4_arch/dshot.h>
 #include <px4_arch/io_timer.h>
-#include <drivers/drv_pwm_output.h>
+#include <drivers/drv_dshot.h>
 
 
 #define MOTOR_PWM_BIT_1				14u
