+++ platforms/nuttx/src/px4/stm/stm32_common/led_pwm/led_pwm.cpp
@@ -56,10 +56,7 @@
 #include <systemlib/err.h>
 #include <systemlib/px4_macros.h>
 
-#include <drivers/drv_pwm_output.h>
 #include <px4_arch/io_timer.h>
-
-
 
 #if defined(BOARD_HAS_LED_PWM)
 
@@ -251,15 +248,15 @@ led_pwm_servo_set(unsigned channel, uint8_t  cvalue)
 
 	return 0;
 }
-unsigned
-led_pwm_servo_get(unsigned channel)
+
+unsigned led_pwm_servo_get(unsigned channel)
 {
 	if (channel >= 3) {
 		return 0;
 	}
 
 	unsigned timer = led_pwm_channels[channel].timer_index;
-	servo_position_t value = 0;
+	uint16_t value = 0;
 
 	/* test timer for validity */
 	if ((led_pwm_timers[timer].base == 0) ||
@@ -289,8 +286,8 @@ led_pwm_servo_get(unsigned channel)
 	unsigned period = led_pwm_timer_get_period(timer);
 	return ((value + 1) * 255 / period);
 }
-int
-led_pwm_servo_init(void)
+
+int led_pwm_servo_init()
 {
 	/* do basic timer initialisation first */
 	for (unsigned i = 0; i < arraySize(led_pwm_timers); i++) {
