+++ src/modules/simulation/pwm_out_sim/PWMSim.cpp
@@ -42,7 +42,6 @@
 #include <px4_platform_common/sem.hpp>
 
 PWMSim::PWMSim(bool hil_mode_enabled) :
-	CDev(PWM_OUTPUT0_DEVICE_PATH),
 	OutputModuleInterface(MODULE_NAME, px4::wq_configurations::hp_default)
 {
 	_mixing_output.setAllDisarmedValues(PWM_SIM_DISARMED_MAGIC);
@@ -51,8 +50,6 @@ PWMSim::PWMSim(bool hil_mode_enabled) :
 	_mixing_output.setAllMaxValues(PWM_SIM_PWM_MAX_MAGIC);
 
 	_mixing_output.setIgnoreLockdown(hil_mode_enabled);
-
-	CDev::init();
 }
 
 void
@@ -65,8 +62,6 @@ PWMSim::Run()
 		exit_and_cleanup();
 		return;
 	}
-
-	SmartLock lock_guard(_lock);
 
 	_mixing_output.update();
 
@@ -119,127 +114,6 @@ PWMSim::updateOutputs(bool stop_motors, uint16_t outputs[MAX_ACTUATORS], unsigne
 	}
 
 	return false;
-}
-
-int
-PWMSim::ioctl(device::file_t *filp, int cmd, unsigned long arg)
-{
-	SmartLock lock_guard(_lock);
-
-	int ret = OK;
-
-	switch (cmd) {
-	case PWM_SERVO_ARM:
-		break;
-
-	case PWM_SERVO_DISARM:
-		break;
-
-	case PWM_SERVO_SET_MIN_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < pwm->channel_count; i++) {
-				if (i < OutputModuleInterface::MAX_ACTUATORS && false) {
-					_mixing_output.minValue(i) = pwm->values[i];
-				}
-			}
-
-			break;
-		}
-
-	case PWM_SERVO_SET_MAX_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < pwm->channel_count; i++) {
-				if (i < OutputModuleInterface::MAX_ACTUATORS && false) {
-					_mixing_output.maxValue(i) = pwm->values[i];
-				}
-			}
-
-			break;
-		}
-
-	case PWM_SERVO_SET_UPDATE_RATE:
-		// PWMSim does not limit the update rate
-		break;
-
-	case PWM_SERVO_SET_SELECT_UPDATE_RATE:
-		break;
-
-	case PWM_SERVO_GET_DEFAULT_UPDATE_RATE:
-		*(uint32_t *)arg = 9999;
-		break;
-
-	case PWM_SERVO_GET_UPDATE_RATE:
-		*(uint32_t *)arg = 9999;
-		break;
-
-	case PWM_SERVO_GET_SELECT_UPDATE_RATE:
-		*(uint32_t *)arg = 0;
-		break;
-
-	case PWM_SERVO_GET_FAILSAFE_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < OutputModuleInterface::MAX_ACTUATORS; i++) {
-				pwm->values[i] = _mixing_output.failsafeValue(i);
-			}
-
-			pwm->channel_count = OutputModuleInterface::MAX_ACTUATORS;
-			break;
-		}
-
-	case PWM_SERVO_GET_DISARMED_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < OutputModuleInterface::MAX_ACTUATORS; i++) {
-				pwm->values[i] = _mixing_output.disarmedValue(i);
-			}
-
-			pwm->channel_count = OutputModuleInterface::MAX_ACTUATORS;
-			break;
-		}
-
-	case PWM_SERVO_GET_MIN_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < OutputModuleInterface::MAX_ACTUATORS; i++) {
-				pwm->values[i] = _mixing_output.minValue(i);
-			}
-
-			pwm->channel_count = OutputModuleInterface::MAX_ACTUATORS;
-			break;
-		}
-
-	case PWM_SERVO_GET_MAX_PWM: {
-			struct pwm_output_values *pwm = (struct pwm_output_values *)arg;
-
-			for (unsigned i = 0; i < OutputModuleInterface::MAX_ACTUATORS; i++) {
-				pwm->values[i] = _mixing_output.maxValue(i);
-			}
-
-			pwm->channel_count = OutputModuleInterface::MAX_ACTUATORS;
-			break;
-		}
-
-	case PWM_SERVO_GET_RATEGROUP(0) ... PWM_SERVO_GET_RATEGROUP(PWM_OUTPUT_MAX_CHANNELS - 1): {
-			// no restrictions on output grouping
-			unsigned channel = cmd - PWM_SERVO_GET_RATEGROUP(0);
-
-			*(uint32_t *)arg = (1 << channel);
-			break;
-		}
-
-	case PWM_SERVO_GET_COUNT:
-		*(unsigned *)arg = OutputModuleInterface::MAX_ACTUATORS;
-		break;
-
-	default:
-		ret = -ENOTTY;
-		break;
-	}
-
-	return ret;
 }
 
 int
