+++ src/modules/vtol_att_control/vtol_att_control_main.h
@@ -50,7 +50,6 @@
 #pragma once
 
 #include <drivers/drv_hrt.h>
-#include <drivers/drv_pwm_output.h>
 #include <lib/geo/geo.h>
 #include <lib/mathlib/mathlib.h>
 #include <lib/perf/perf_counter.h>
